+++ common/src/main/java/net/mca/Config.java
@@ -139,6 +139,7 @@ public final class Config implements Serializable {
     public boolean allowDestinyCommandOnce = true;
     public boolean allowDestinyCommandMoreThanOnce = false;
     public boolean allowDestinyTeleportation = true;
+    public boolean enablePlayerShaders = true;
     public boolean enableVillagerPlayerModel = true;
     public boolean forceVillagerPlayerModel = false;
     public boolean allowLimitedPlayerEditor = true;
@@ -216,6 +217,10 @@ public final class Config implements Serializable {
     );
 
     public Map<String, String> professionConversionsMap = Map.of();
+
+    public Map<String, String> shaderLocationsMap = Map.of(
+            "color_blind", "minecraft:shaders/post/desaturate.json"
+    );
 
     public Map<String, String> playerRendererBlacklist = Map.of(
             "firstpersonmod", "arms",